pyproject.toml
[tool.poetry]
name = "technocracy_practice_test"
version = "0.1.0"
description = ""
authors = ["eale <RUSLIKPVP@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
python-dotenv = "^0.20.0"
djangorestframework-simplejwt = "^5.2.0"
pytest-cov = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "notes.settings"
filterwarnings = [
    "error",
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]
python_files = ['test_*.py']
addopts = ['-vv', '--cov', '-p', 'no:warnings']
